fixed type retention

+++ internal/domain/models/invoice.go
@@ -8,18 +8,18 @@ import (
 )
 
 type Invoice struct {
-	InvoiceID      uuid.UUID      `db:"invoice_id"`
-	ProjectID      uuid.UUID      `db:"project_id"`
-	PeriodID       uuid.UUID      `db:"period_id"`
-	InvoiceDate    sql.NullTime   `db:"invoice_date"`
-	PaymentDueDate sql.NullTime   `db:"payment_due_date"`
-	PaidDate       sql.NullTime   `db:"paid_date"`
-	PaymentTerm    sql.NullString `db:"payment_term"`
-	Remarks        sql.NullString `db:"remarks"`
-	Status         sql.NullString `db:"status"`
-	CreatedAt      time.Time      `db:"created_at"`
-	UpdatedAt      sql.NullTime   `db:"updated_at"`
-	Retention      sql.NullInt32  `db:"retention"`
+	InvoiceID      uuid.UUID       `db:"invoice_id"`
+	ProjectID      uuid.UUID       `db:"project_id"`
+	PeriodID       uuid.UUID       `db:"period_id"`
+	InvoiceDate    sql.NullTime    `db:"invoice_date"`
+	PaymentDueDate sql.NullTime    `db:"payment_due_date"`
+	PaidDate       sql.NullTime    `db:"paid_date"`
+	PaymentTerm    sql.NullString  `db:"payment_term"`
+	Remarks        sql.NullString  `db:"remarks"`
+	Status         sql.NullString  `db:"status"`
+	CreatedAt      time.Time       `db:"created_at"`
+	UpdatedAt      sql.NullTime    `db:"updated_at"`
+	Retention      sql.NullFloat64 `db:"retention"`
 
 	// Related data (not stored in database)
 	Period Period `db:"-"`
